refactor: template generator refactor

+++ packages/fx-core/src/component/generator/generator.ts
@@ -17,13 +17,11 @@ import {
   sampleDefaultTimeoutInMs,
 } from "./constant";
 import {
-  CancelDownloading,
   DownloadSampleApiLimitError,
   DownloadSampleNetworkError,
   FetchSampleInfoError,
   TemplateNotFoundError,
   TemplateZipFallbackError,
-  UnzipError,
 } from "./error";
 import {
   SampleActionSeq,
@@ -164,7 +162,6 @@ export class Generator {
         await action.run(context);
         await context.onActionEnd?.(action, context);
       } catch (e) {
-        if (e instanceof BaseComponentInnerError) throw e.toFxError();
         if (e instanceof Error) await context.onActionError(action, context, e);
       }
     }
@@ -177,20 +174,18 @@ export function templateDefaultOnActionError(
   error: Error
 ): Promise<void> {
   switch (action.name) {
-    case GeneratorActionName.FetchUrlForHotfixOnly:
-    case GeneratorActionName.FetchZipFromUrl:
-      context.cancelDownloading = true;
-      if (!(error instanceof CancelDownloading)) {
-        context.logProvider.info(error.message);
-        context.logProvider.info(LogMessages.getTemplateFromLocal);
-      }
+    case GeneratorActionName.RemoteTemplate:
+      context.fallback = true;
+      context.logProvider.info(error.message);
+      context.logProvider.info(LogMessages.getTemplateFromLocal);
       break;
-    case GeneratorActionName.FetchTemplateZipFromLocal:
-      context.logProvider.error(error.message);
-      return Promise.reject(new TemplateZipFallbackError().toFxError());
-    case GeneratorActionName.Unzip:
-      context.logProvider.error(error.message);
-      return Promise.reject(new UnzipError().toFxError());
+    case GeneratorActionName.LocalTemplate:
+      if (error instanceof BaseComponentInnerError) {
+        return Promise.reject(error.toFxError());
+      } else {
+        context.logProvider.error(error.message);
+        return Promise.reject(new TemplateZipFallbackError().toFxError());
+      }
     default:
       return Promise.reject(new Error(error.message));
   }
@@ -203,6 +198,7 @@ export async function sampleDefaultOnActionError(
   error: Error
 ): Promise<void> {
   context.logProvider.error(error.message);
+  if (error instanceof BaseComponentInnerError) throw error.toFxError();
   if (await fs.pathExists(context.destination)) {
     await fs.rm(context.destination, { recursive: true });
   }
